requirements.txt
python-dotenv==1.0.1
motor==3.7.1
aiohttp==3.10.5
aiofiles==23.2.1
python-docx==1.2.0
aiogram==3.20.0.post0
g4f==0.5.5.5
torch==2.5.1
scipy==1.16.0
transformers==4.46.2
selenium==4.27.1
pandas==2.3.0